🔧 fix: 修复跨平台编译警告（指针↔整数、浮点转换、整数截断）
- eh_types.h: 新增 eh_isinf/eh_isnan 宏封装编译器 builtin
- eh_rbtree.h: parent_node_t 从 unsigned long 改为 uintptr_t
- eh_error.h: (long) 强转改为 (intptr_t) 保证 LLP64/LP64 兼容
- eh_list.h: 用 offsetof() 替代手动偏移量计算
- eh_mem_pool.c: 补充 #include <stdint.h>，unsigned long → uintptr_t
- eh_formatio.c: isinf/isnan 改用 eh_isinf/eh_isnan 避免 MinGW 宏浮点警告
- eh_core.c: uintptr_t → long 显式转换防止截断警告
- eh_internal.h: eh_init_fini_array_len 改为 uintptr_t
- platform.c: LONGLONG → uint64_t 显式转换消除符号警告
- module_section.c: DWORD 加法结果显式转为 long

+++ src/eh_core.c
@@ -368,7 +368,7 @@ static int interior_init(void){
     eh->dispatch_cnt = 0;
     eh->idle_time = 0;
     eh->eh_init_fini_array = (struct eh_module*)eh_module_section_begin();
-    eh->eh_init_fini_array_len = ((struct eh_module*)eh_module_section_end() - (struct eh_module*)eh_module_section_begin());
+    eh->eh_init_fini_array_len = (uintptr_t)((struct eh_module*)eh_module_section_end() - (struct eh_module*)eh_module_section_begin());
     return 0;
 }
 
@@ -381,7 +381,7 @@ static int  module_group_init(void){
     struct eh_module  *eh_init_fini_array;
     long i,len;
     int ret = 0;
-    len = eh->eh_init_fini_array_len;
+    len = (long)eh->eh_init_fini_array_len;
     eh_init_fini_array = eh->eh_init_fini_array;
     for(i=0;i<len;i++){
         if(eh_init_fini_array[i].init){
@@ -406,7 +406,7 @@ static void module_group_exit(void){
     eh_t *eh = eh_get_global_handle();
     struct eh_module  *eh_init_fini_array;
     long i;
-    i = eh->eh_init_fini_array_len;
+    i = (long)eh->eh_init_fini_array_len;
     eh_init_fini_array = eh->eh_init_fini_array;
 
     for(i=i-1;i>=0;i--){
